Добавил обработчики: - клика на карточку фильма для открытия попапа с ифнормацией - на кнопку ESC и крестик для закрытия попап окна с информацией
