+++ src/presenter/films-presenter.js
@@ -5,6 +5,7 @@ import FilmsListTitleView from '../view/films-list-title-view';
 import FilmsContainerView from '../view/films-container-view';
 import FilmCardView from '../view/film-card-view';
 import ShowMoreButtonView from '../view/show-more-button-view';
+import FilmDetailsView from '../view/film-details-view';
 
 export default class FilmsPresenter {
   #filmsContent = new FilmsContentView();
@@ -33,6 +34,31 @@ export default class FilmsPresenter {
 
   #renderFilm = (film) => {
     const filmComponent = new FilmCardView(film);
+    const filmDetailsComponent = new FilmDetailsView(film);
+
+    const closingFilmDetails = (event) => {
+      event.preventDefault();
+      this.#filmsList.element.removeChild(filmDetailsComponent.element);
+      document.body.classList.remove('hide-overflow');
+      document.removeEventListener('keydown', keyDownClosingHandler);
+    };
+
+    const clickClosingHandler = closingFilmDetails;
+
+    function keyDownClosingHandler(event) {
+      if(event.key === 'Escape'){
+        closingFilmDetails(event);
+      }
+    }
+
+    filmComponent.link.addEventListener('click', (event) => {
+      event.preventDefault();
+      this.#filmsList.element.appendChild(filmDetailsComponent.element);
+      document.body.classList.add('hide-overflow');
+      document.addEventListener('keydown', keyDownClosingHandler);
+    });
+
+    filmDetailsComponent.closeButton.addEventListener('click', clickClosingHandler);
 
     render(filmComponent, this.#filmsContainer.element);
   };
