+++ src/view/film-details-view.js
@@ -1,0 +1,208 @@
+import {createElement} from '../render';
+import {transformFilmReleaseDateFull, transformFilmRuntime} from '../utils/utils';
+
+const createFilmDetailsTemplate = (film) => {
+  const { title, alternativeTitle, totalRating, ageRating, description, poster, runtime, release, genre, actors, director, writers } = film.filmInfo;
+  const { watchlist, alreadyWatched, favorite } = film.userDetails;
+
+  const date = release.date !== null
+    ? transformFilmReleaseDateFull(release.date)
+    : '-';
+
+  const timeDuration = runtime !== null
+    ? transformFilmRuntime(runtime)
+    : '-';
+
+  const setClassActiveControls = (status) => status ? 'film-details__control-button--active' : '';
+
+  const setGenres = () => genre.map((text) => `<span className="film-details__genre">${text}</span>`).join(' ');
+
+  return (
+    `<section class="film-details">
+        <form class="film-details__inner" action="" method="get">
+          <div class="film-details__top-container">
+            <div class="film-details__close">
+              <button class="film-details__close-btn" type="button">close</button>
+            </div>
+            <div class="film-details__info-wrap">
+              <div class="film-details__poster">
+                <img class="film-details__poster-img" src="${poster}" alt="">
+
+                <p class="film-details__age">+${ageRating}</p>
+              </div>
+
+              <div class="film-details__info">
+                <div class="film-details__info-head">
+                  <div class="film-details__title-wrap">
+                    <h3 class="film-details__title">${title}</h3>
+                    <p class="film-details__title-original">${alternativeTitle}</p>
+                  </div>
+
+                  <div class="film-details__rating">
+                    <p class="film-details__total-rating">${totalRating}</p>
+                  </div>
+                </div>
+
+                <table class="film-details__table">
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Director</td>
+                    <td class="film-details__cell">${director}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Writers</td>
+                    <td class="film-details__cell">${writers.join(', ')}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Actors</td>
+                    <td class="film-details__cell">${actors.join(', ')}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Release Date</td>
+                    <td class="film-details__cell">${date}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Runtime</td>
+                    <td class="film-details__cell">${timeDuration}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Country</td>
+                    <td class="film-details__cell">${release.releaseCountry}</td>
+                  </tr>
+                  <tr class="film-details__row">
+                    <td class="film-details__term">Genres</td>
+                    <td class="film-details__cell">${setGenres()}</tr>
+                </table>
+
+                <p class="film-details__film-description">${description}</p>
+              </div>
+            </div>
+
+            <section class="film-details__controls">
+              <button type="button" class="film-details__control-button film-details__control-button--watchlist ${setClassActiveControls(watchlist)}" id="watchlist" name="watchlist">Add to watchlist</button>
+              <button type="button" class="film-details__control-button film-details__control-button--watched ${setClassActiveControls(alreadyWatched)}" id="watched" name="watched">Already watched</button>
+              <button type="button" class="film-details__control-button film-details__control-button--favorite ${setClassActiveControls(favorite)}" id="favorite" name="favorite">Add to favorites</button>
+            </section>
+          </div>
+          <div class="film-details__bottom-container">
+          <section class="film-details__comments-wrap">
+            <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">4</span></h3>
+
+            <ul class="film-details__comments-list">
+              <li class="film-details__comment">
+                <span class="film-details__comment-emoji">
+                  <img src="./images/emoji/smile.png" width="55" height="55" alt="emoji-smile">
+                </span>
+                <div>
+                  <p class="film-details__comment-text">Interesting setting and a good cast</p>
+                  <p class="film-details__comment-info">
+                    <span class="film-details__comment-author">Tim Macoveev</span>
+                    <span class="film-details__comment-day">2019/12/31 23:59</span>
+                    <button class="film-details__comment-delete">Delete</button>
+                  </p>
+                </div>
+              </li>
+              <li class="film-details__comment">
+                <span class="film-details__comment-emoji">
+                  <img src="./images/emoji/sleeping.png" width="55" height="55" alt="emoji-sleeping">
+                </span>
+                <div>
+                  <p class="film-details__comment-text">Booooooooooring</p>
+                  <p class="film-details__comment-info">
+                    <span class="film-details__comment-author">John Doe</span>
+                    <span class="film-details__comment-day">2 days ago</span>
+                    <button class="film-details__comment-delete">Delete</button>
+                  </p>
+                </div>
+              </li>
+              <li class="film-details__comment">
+                <span class="film-details__comment-emoji">
+                  <img src="./images/emoji/puke.png" width="55" height="55" alt="emoji-puke">
+                </span>
+                <div>
+                  <p class="film-details__comment-text">Very very old. Meh</p>
+                  <p class="film-details__comment-info">
+                    <span class="film-details__comment-author">John Doe</span>
+                    <span class="film-details__comment-day">2 days ago</span>
+                    <button class="film-details__comment-delete">Delete</button>
+                  </p>
+                </div>
+              </li>
+              <li class="film-details__comment">
+                <span class="film-details__comment-emoji">
+                  <img src="./images/emoji/angry.png" width="55" height="55" alt="emoji-angry">
+                </span>
+                <div>
+                  <p class="film-details__comment-text">Almost two hours? Seriously?</p>
+                  <p class="film-details__comment-info">
+                    <span class="film-details__comment-author">John Doe</span>
+                    <span class="film-details__comment-day">Today</span>
+                    <button class="film-details__comment-delete">Delete</button>
+                  </p>
+                </div>
+              </li>
+            </ul>
+
+            <div class="film-details__new-comment">
+              <div class="film-details__add-emoji-label"></div>
+
+              <label class="film-details__comment-label">
+                <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
+              </label>
+
+              <div class="film-details__emoji-list">
+                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
+                <label class="film-details__emoji-label" for="emoji-smile">
+                  <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
+                </label>
+
+                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
+                <label class="film-details__emoji-label" for="emoji-sleeping">
+                  <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
+                </label>
+
+                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
+                <label class="film-details__emoji-label" for="emoji-puke">
+                  <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
+                </label>
+
+                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
+                <label class="film-details__emoji-label" for="emoji-angry">
+                  <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
+                </label>
+              </div>
+            </div>
+          </section>
+        </div>
+      </form>
+    </section>`
+  );
+};
+
+export default class FilmDetailsView {
+  #element = null;
+  #film = null;
+
+  constructor(film) {
+    this.#film = film;
+  }
+
+  get template() {
+    return createFilmDetailsTemplate(this.#film);
+  }
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+
+    return this.#element;
+  }
+
+  get closeButton() {
+    return this.element.querySelector('.film-details__close-btn');
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
